udpates

+++ itemTransfer/models.py
@@ -6,12 +6,12 @@ class ItemType(BaseModel):
     itemtransferId: Optional[str] = None  
     itemCode: Optional[List[str]] = None
     itemName: Optional[List[str]] = None
-    reqQty: Optional[List[float]] = None    
+    reqQty: Optional[List[int]] = None    
     uom: Optional[List[str]] = None
     price: Optional[List[int]] = None
     totalPrice: Optional[float] = None 
-    receivedQty: Optional[List[float]] = None
-    sendQty: Optional[List[float]] = None
+    receivedQty: Optional[List[int]] = None
+    sendQty: Optional[List[int]] = None
     fromBranch: Optional[str] = None
     toBranch: Optional[str] = None
     fromLoginId: Optional[str] = None
@@ -26,13 +26,13 @@ class ItemType(BaseModel):
 class ItemTypePost(BaseModel):
     itemCode: Optional[List[str]] = None
     itemName: Optional[List[str]] = None
-    reqQty: Optional[List[float]] = None
+    reqQty: Optional[List[int]] = None
     uom: Optional[List[str]] = None
     price: Optional[List[int]] = None
     totalPrice: Optional[float] = None
     approvedQty: Optional[List[int]] = None
-    receivedQty: Optional[List[float]] = None
-    sendQty: Optional[List[float]] = None
+    receivedQty: Optional[List[int]] = None
+    sendQty: Optional[List[int]] = None
     fromBranch: Optional[str] = None
     toBranch: Optional[str] = None
     fromLoginId: Optional[str] = None
